Aplicativo/app/requirements.txt
Kivy
plyer
pillow
jnius
opencv
numpy

